resolvers,typeDefs,mutations,queries

+++ server/schemas/resolvers.js
@@ -23,35 +23,35 @@ const resolvers = {
   Query: {
     // Return the current logged-in employee's info
     me: async (_, __, context) => {
-      /* if (context.user) { */
+      if (context.user) {
         return await User.findById(context.user._id);
-      // }
+      }
       throw AuthenticationError;
     },
 
     // Get all employees, optionally filter by name or ID
     employees: async (_, __, context) => {
-      // if (context.user) {
+      if (context.user) {
         return await Employee.find();
-      // }
+      }
       throw AuthenticationError;
     },
 
     // Get a single employee by ID
-    employee: async (_, { userId }) => {
-      // if (context.user) {
-        return await Employee.findOne({user:{userId}}).populate("TimeOffRequest");
-      // }
+    employee: async (_, args,context) => {
+      if (context.user) {
+        return await Employee.findOne({user:context.user._id})
+      }
       throw AuthenticationError;
     },
 
     // Get all time off requests
-    timeOffRequests: async (_, __, context) => {
-      // if (context.user) {
+    /* timeOffRequests: async (_, __, context) => {
+      if (context.user) {
         return await TimeOffRequest.find();
-      // }
+      }
       throw AuthenticationError;
-    },
+    }, */
   },
 
   Mutation: {
@@ -76,25 +76,24 @@ const resolvers = {
         throw AuthenticationError;
       }
 
-      const newUser = await User.create({
+      const user = await User.create({
         username,
         password,
+        email
       });
 
       await Employee.findOneAndUpdate(
         {_id:employee._id},
-        {$set:{user:newUser}},
+        {$set:{user:user}},
         {new:true}
       )
-      /* employee.user.set(newUser);
-      await employee.save(); */
-      const token = signToken(newUser);
-      return { token, newUser };
+      const token = signToken(user);
+      return { token, user };
     },
 
     // Add a new employee
     addEmployee: async (_, {employee}, context) => {
-      // if (context.user) {
+      /* if (context.user) { */
         const newEmployee = await Employee.create(employee);
         return newEmployee;
       // }
@@ -102,10 +101,10 @@ const resolvers = {
     },
 
     // Update an employee's details
-    updateEmployee: async (_, args) => {
-      // if (context.user) {
+    updateEmployee: async (_, args,context) => {
+      if (context.user) {
         return await Employee.findByIdAndUpdate(
-          { _id: empId },
+          { _id: args.empId },
           {
             firstName: args.firstName,
             lastName: args.lastName,
@@ -115,23 +114,23 @@ const resolvers = {
           },
           { new: true }
         );
-      // }
+      }
       throw AuthenticationError;
     },
 
     // Terminate an employee (only accessible to admin)
     terminateEmployee: async (_, { userId }, context) => {
-      // if (context.user.isAdmin) {
+      if (context.user/* .isAdmin */) {
         const user = await User.findOneAndDelete({ _id: userId });
         return await Employee.findOneAndDelete(user);
-      // }
+      }
       throw AuthenticationError;
     },
 
     // Create a time off request for an employee
-    createTimeOffRequest: async (_, { empId, startDate, endDate }) => {
-      // if (context.user) {
-        const request = await TimeOffRequest.create({
+    createTimeOffRequest: async (_, { empId, startDate, endDate },context) => {
+      if (context.user) {
+        /* const request = await TimeOffRequest.create({
           startDate,
           endDate,
         });
@@ -140,20 +139,32 @@ const resolvers = {
           { $addToSet: { timeOffRequests: request } },
           { new: true }
         );
-        return request;
-      // }
+        return request; */
+
+        return await Employee.findOneAndUpdate(
+          {_id:empId},
+          {$addToSet:{timeOffRequests:{startDate,endDate}}},
+          {new:true}
+        )
+      }
       throw AuthenticationError;
     },
 
     // Update the status of a time off request (pending, approved, denied)
-    updateTimeOffRequestStatus: async (_, { requestId, status }) => {
-      // if (context.user.isAdmin) {
-        return await TimeOffRequest.findByIdAndUpdate(
+    updateTimeOffRequestStatus: async (_, { empId,requestId, status },context) => {
+      if (context.user/* .isAdmin */) {
+        /* return await TimeOffRequest.findByIdAndUpdate(
           requestId,
           { status },
           { new: true }
-        );
-      // }
+        ); */
+        const employee = await Employee.findOne({_id:empId})
+        employee.timeOffRequests.id(requestId).status = status;
+        employee.markModified('timeOffRequests')
+        employee.save();
+        return employee;
+
+      }
       throw AuthenticationError;
     },
   },
